feat: move formatting to extra directory, add regex for only one occurrence of punctuation, replace , with ., update error handling
